effects: give sandbox and catch a home that is not node's
`effects/node` declared both because node was the only host that ran
them, which is not the criterion: an operation belongs to the layer of
whoever implements it, and a browser interpreter implements these two —
the shared proof traversal performs exactly `sandbox` and `catch`. Until
they move, a browser cannot reach the logic without importing a module
named after a different host.

`effects/common` is the destination, one directory rather than the four
`node-module-layering.md` proposed. The goal that settles it is the
project's direction — less JavaScript, the logic in FunctionalScript,
every host talking to it through generic effects — under which the
meaningful axis is *which host implements this*, not *what concern is
it*. The issue's table and tasks are rewritten to that, with the reason
recorded rather than the naming quietly changed.

A pure move: `effects/node` re-exports both, which is a live coupling
and not a shim, since `NodeOp` unions them. The runtime-agnostic callers
in `emergent_testing` now name `effects/common` directly, which is the
part that pays: what a second host must implement is one import, not an
archaeology exercise.

`Await` stays put deliberately — it belongs to the registration path no
browser runs, so it moves on the layering argument alone and is not
urgent. `All` likewise: the run path has performed none since the
traversal went sequential.

The new module's proof interprets both through a handler map keyed by
`'sandbox'` and `'catch'`, because those strings are the whole agreement
between the two sides — a renamed command is a type error nowhere and an
unclaimed command at runtime.
