+++ changelog/unreleased/1791.md
@@ -1,0 +1,3 @@
+- `effects`: new `effects/common`, for operations more than one host
+  implements. `Sandbox` and `Catch` move there from `effects/node`, which
+  re-exports both, so existing imports keep working
