+++ fjs/effects/node/types.ts
@@ -15,6 +15,7 @@ import type {
     Operation, ToAsyncOperationMap,
 } from '../types.ts'
 import type { List } from '../list/types.ts'
+import type { Catch, Sandbox, SandboxResult } from '../common/types.ts'
 
 /**
  * The vocabulary every operation is declared in — how a runner reports that it
@@ -25,6 +26,15 @@ import type { List } from '../list/types.ts'
  * on reading as one vocabulary rather than two.
  */
 export type { IoChannel, IoError, IoErrorInfo, IoResult, OpResult }
+
+/**
+ * `Sandbox` and `Catch` are declared in [`../common`](../common/types.ts) —
+ * they have a second implementer, and an operation belongs to the layer of
+ * whoever implements it. They are re-exported here because `NodeOp` unions
+ * them and dozens of signatures name them through this module; that makes this
+ * a live coupling rather than a shim.
+ */
+export type { Catch, Sandbox, SandboxResult }
 
 // all
 
@@ -269,25 +279,6 @@ export type Now = readonly['now', () => OpResult<number>]
 // sandbox
 
 /**
- * The outcome of a `Sandbox` operation.
- *
- * `result` carries either `['ok', value]` or `['error', thrown]`. `duration`
- * is a floating-point millisecond count with up to microsecond precision,
- * matching `performance.now()` directly. Additional fields (allocated memory,
- * max stack depth, coverage) may be added in future without breaking consumers.
- */
-export type SandboxResult<T> = {
-    readonly result: Result<T, unknown>
-    /**
-     * Elapsed time in milliseconds (microsecond precision via `performance.now()`).
-     * The virtual runner returns `0` for deterministic tests.
-     */
-    readonly duration: number
-}
-
-export type Sandbox = readonly['sandbox', <T>(f: () => T) => OpResult<SandboxResult<T>>]
-
-/**
  * Resolves the return value of a test function inside the effect runner.
  * If `p` is a real `Promise`, it is awaited and rejections propagate as
  * throws. If `p` is any other value it is returned as-is. Plain thenables
@@ -295,29 +286,6 @@ export type Sandbox = readonly['sandbox', <T>(f: () => T) => OpResult<SandboxRes
  * treated as ordinary values — not awaited. See `fjs/dev/tf/README.md`.
  */
 export type Await = readonly['await', (p: unknown) => OpResult<readonly[unknown]>]
-
-// catch
-
-/**
- * Runs a pure thunk and answers what it did: its value, or the value it threw.
- *
- * It sits beside {@link Sandbox} and is deliberately *not* it. `sandbox` carries
- * a clock and, in the virtual runner, a fixture convention — its handler is a
- * pass-through whose thunk is expected to answer a {@link SandboxResult}
- * directly, because `../virtual` is `.f.mjs` and FunctionalScript has no
- * `try`/`catch` to implement a real one with. Routing a tree walk through
- * `sandbox` would hand that handler a thunk answering something else entirely.
- *
- * This one carries neither, so every runner implements it truthfully: the real
- * Node runner and a browser interpreter with `tryCatch`, and the virtual runner
- * with `ok(ok(f()))` — a pure runner still cannot catch, so a hostile fixture
- * still panics there, which is the same bargain `sandbox` already makes.
- *
- * It exists because reading a *user* value is an operation, not pure logic: the
- * proof traversal enumerates values a test returned, and an enumerable getter or
- * a proxy trap in one of them is a failure of that test rather than of the run.
- */
-export type Catch = readonly['catch', <T>(f: () => T) => OpResult<Result<T, unknown>>]
 
 // Test registration
